Move data manager concerns
