+++ src/main/data-managers/DeviceManager.js
